Reformats GBIF data into dynamicSDM data frame

Function converts GBIF occurrence records into the format required for dynamicSDM functions.

convert_gbif(gbif.df)

Arguments

gbif.df

a data frame, the direct output from GBIF occurrence record download.

Value

Returns data frame correctly formatted for input into dynamicSDM functions.

Details

For most dynamicSDM functions, an occurrence data frame with record co-ordinate columns labelled "x" and "y" with numeric columns for record "day", "month" and "year" are required. This functions takes the input data frame and returns a reformatted data frame suitable for direct input into dynamicSDM functions.

Examples

data(sample_occ_data)
converted <- convert_gbif(sample_occ_data)

if (FALSE) {

gbif.df <- data.frame(rgbif::occ_search(scientificName = "Quelea_quelea")$data)

occ.data <- convert_gbif(gbif.df)
}