Combine extracted explanatory variable data for occurrence records into single data frame.

Combines the split output files from functions extract_dynamic_coords() and extract_buffered_coords() into single data frame containing all occurrence records and explanatory variables.

extract_coords_combine(varnames, local.directory)

Arguments

varnames

a character string, the unique names for each explanatory variable.

local.directory

a character string or vector, the path to local directory or directories to read extracted explanatory data frames from.

Value

Returns a data frame containing all occurrence records with associated explanatory variable data.

Details

When functions extract_dynamic_coords() and extract_buffered_coords() have been used to extract dynamic explanatory variables for occurrence records, the output for individual records and each variable will be split into separate “csv” files.

This function reads in these files and combines data into a single data frame containing each occurrence records and associated explanatory data from each variable.

To prevent error, the “csv” files must be uniquely named within the folder(s) and include an exact character match for the varnames provided. All “csv” files matching the varnames should have the same number and names of columns. This is the default output ofextract_dynamic_coords() and extract_buffered_coords().

Examples

if (FALSE) { # googledrive::drive_has_token()
data(sample_filt_data)
# \dontshow{
sample_filt_data<-sample_filt_data[1:2,]
# }

dynamicSDM::extract_dynamic_coords(
 occ.data = sample_filt_data,
 datasetname = "UCSB-CHG/CHIRPS/DAILY",
 bandname = "precipitation",
 spatial.res.metres = 10000,
 GEE.math.fun = "sum",
 temporal.direction = "prior",
 temporal.res = 56,
 save.method = "split",
 varname = "eightweekprec",
 save.directory = tempdir()
)
dynamicSDM::extract_dynamic_coords(
 occ.data = sample_filt_data,
 datasetname = "UCSB-CHG/CHIRPS/DAILY",
 bandname = "precipitation",
 spatial.res.metres = 10000,
 GEE.math.fun = "sum",
 temporal.direction = "prior",
 temporal.res = 364,
 save.method = "combined",
 varname = "annualweekprec",
 save.directory = tempdir()
)

extract_coords_combine(varnames = c("eightweekprec", "annualweekprec"), local.directory = tempdir())
}