Sample species occurrence records with associated dynamic explanatory variables

A dataset containing a sample of the bird species, the red-billed quelea (Quelea quelea), distribution records from between 2002-2019 (GBIF 2021, GBIF 2022); generated pseudo-absence records, and associated extracted dynamic explanatory variables. The variables are as follows:

sample_explan_data

Format

A data frame with 330 rows and 17 variables:

x

species occurrence record longitude.

y

species occurrence record latitude.

year

species occurrence record year.

month

species occurrence record month.

day

species occurrence record day.

decimalLatitude

species occurrence record latitude.

decimalLongitude

species occurrence record longitude.

occurrenceStatus

species presence or absence character.

source

source of occurrence or pseudo-absence data point.

species

name of species occurrence records belong to name

SAMP_EFFORT

total number of avian e-Bird sampling events within spatiotemporal buffer of occurrence record location and dates.

REL_SAMP_EFFORT

proportion of total number of avian e-Bird sampling events within spatiotemporal buffer of occurrence record location and dates relative to other records

unique.ID.DYN

unique id value assigned when extracting dynamic explanatory variable data

eight_sum_prec

sum Climate Hazards Group InfraRed Precipitation With Station Data (Funk et al., 2016) total daily precipitation at record co-ordinate across 52-weeks prior to record date (mm).

grass_crop_percentage

total number of MODIS Land Cover Type Yearly 500m (Friedl & Sulla-Menashe, 2019) "cereal cropland" and "grassland" cells in surrounding area of record co-ordinate in record year.

year_sum_prec

sum Climate Hazards Group InfraRed Precipitation With Station Data (CHIRPS Daily) total daily precipitation at record co-ordinate across 52-weeks prior to record date (mm).

presence.absence

binary species presence or absence at record location and date.

References

Friedl, M., Sulla-Menashe, D. (2019). MCD12Q1 MODIS/Terra+Aqua Land Cover Type Yearly L3 Global 500m SIN Grid V006. NASA EOSDIS Land Processes DAAC. Accessed 2022-11-24 from https://doi.org/10.5067/MODIS/MCD12Q1.006

Funk, Chris, Pete Peterson, Martin Landsfeld, Diego Pedreros, James Verdin, Shraddhanand Shukla, Gregory Husak, James Rowland, Laura Harrison, Andrew Hoell & Joel Michaelsen. "The climate hazards infrared precipitation with stations-a new environmental record for monitoring extremes". Scientific Data 2, 150066. doi:10.1038/sdata.2015.66 2015.

GBIF.org (12 July 2021) GBIF Occurrence Download https://doi.org/10.15468/dl.ppcu6q

GBIF.org (25 July 2022) GBIF Occurrence Download https://doi.org/10.15468/dl.k2kftv