Test for spatial and temporal autocorrelation in species distribution model explanatory data.

Function performs statistical tests to assess spatial and temporal autocorrelation in given explanatory variable data.

spatiotemp_autocorr(occ.data, varname, temporal.level, plot = F)

Arguments

occ.data

a data frame, with columns for occurrence record co-ordinates and dates with column names as follows; record longitude as "x", latitude as "y", year as "year", month as "month", and day as "day" and associated explanatory data.

varname

a character string or vector, the name(s) of the columns within occ.data containing data to test for autocorrelation.

temporal.level

a character string or vector, the time step(s) to test for temporal autocorrelation at. One or multiple of day or month, year. Can be abbreviated.

plot

a logical indicating whether to generate plot of temporal autocorrelation. See details for plot description. Default = F.

Value

Returns a list of temporal and spatial autocorrelation test results for each variable.

Details

To test for temporal autocorrelation, the function first calculates the average value across records for each time step (temporal.level). The correlation between the average value at one time point (t) and the value at the previous time point (t-1) is calculated and plotted (if plot = T) A significant relationship between values at consecutive data points indicates temporal autocorrelation is present.

To test for spatial autocorrelation, the function calculates a distance matrix between all record co-ordinates. Moran’s I statistical test is calculated to test whether points closer in space have more similar values than those more distant from each other. Please note that NA values are removed before Moran's I calculation.

As the spatial autocorrelation calculation involves computation of a distance matrix between all occurrence records. To reduce computation time, it is recommended that a sample of large occurrence datasets are input.

Examples

data("sample_explan_data")
spatiotemp_autocorr(sample_explan_data,
                   varname = c("year_sum_prec","eight_sum_prec"),
                   temporal.level = c("year","month","day"))
#> $year_sum_prec
#> $year_sum_prec$Temporal_autocorrelation
#> $year_sum_prec$Temporal_autocorrelation$year
#> 
#> 	Pearson's product-moment correlation
#> 
#> data:  first_obs and second_obs
#> t = 0.71217, df = 14, p-value = 0.4881
#> alternative hypothesis: true correlation is not equal to 0
#> 95 percent confidence interval:
#>  -0.3402656  0.6247751
#> sample estimates:
#>       cor 
#> 0.1869775 
#> 
#> 
#> $year_sum_prec$Temporal_autocorrelation$month
#> 
#> 	Pearson's product-moment correlation
#> 
#> data:  first_obs and second_obs
#> t = -1.2414, df = 9, p-value = 0.2459
#> alternative hypothesis: true correlation is not equal to 0
#> 95 percent confidence interval:
#>  -0.7989694  0.2822709
#> sample estimates:
#>        cor 
#> -0.3823456 
#> 
#> 
#> $year_sum_prec$Temporal_autocorrelation$day
#> 
#> 	Pearson's product-moment correlation
#> 
#> data:  first_obs and second_obs
#> t = 0.68107, df = 28, p-value = 0.5014
#> alternative hypothesis: true correlation is not equal to 0
#> 95 percent confidence interval:
#>  -0.2438259  0.4664728
#> sample estimates:
#>       cor 
#> 0.1276571 
#> 
#> 
#> 
#> $year_sum_prec$Spatial_autocorrelation
#>      observed     expected          sd p.value
#> 1 -0.09107117 -0.003039514 0.001845888       0
#> 
#> 
#> $eight_sum_prec
#> $eight_sum_prec$Temporal_autocorrelation
#> $eight_sum_prec$Temporal_autocorrelation$year
#> 
#> 	Pearson's product-moment correlation
#> 
#> data:  first_obs and second_obs
#> t = 1.9975, df = 14, p-value = 0.06558
#> alternative hypothesis: true correlation is not equal to 0
#> 95 percent confidence interval:
#>  -0.03229165  0.78370005
#> sample estimates:
#>       cor 
#> 0.4709523 
#> 
#> 
#> $eight_sum_prec$Temporal_autocorrelation$month
#> 
#> 	Pearson's product-moment correlation
#> 
#> data:  first_obs and second_obs
#> t = 4.0783, df = 9, p-value = 0.002765
#> alternative hypothesis: true correlation is not equal to 0
#> 95 percent confidence interval:
#>  0.3979590 0.9475389
#> sample estimates:
#>       cor 
#> 0.8055327 
#> 
#> 
#> $eight_sum_prec$Temporal_autocorrelation$day
#> 
#> 	Pearson's product-moment correlation
#> 
#> data:  first_obs and second_obs
#> t = -0.61025, df = 28, p-value = 0.5466
#> alternative hypothesis: true correlation is not equal to 0
#> 95 percent confidence interval:
#>  -0.4560139  0.2562806
#> sample estimates:
#>        cor 
#> -0.1145665 
#> 
#> 
#> 
#> $eight_sum_prec$Spatial_autocorrelation
#>      observed     expected          sd      p.value
#> 1 -0.01417411 -0.003039514 0.001844916 1.586784e-09
#> 
#>