Test for spatial and temporal bias in species occurrence records

Generates plots for visual assessment of spatial and temporal biases in occurrence records. Tests whether the spatiotemporal distribution of records is significantly different from the distribution from random sampling.

spatiotemp_bias(
  occ.data,
  temporal.level,
  plot = F,
  spatial.method = "simple",
  centroid,
  radius,
  prj = "+proj=longlat +datum=WGS84"
)

Arguments

occ.data

a data frame, with columns for occurrence record co-ordinates and dates with column names as follows; record longitude as "x", latitude as "y", year as "year", month as "month", and day as "day".

temporal.level

a character string or vector, the time step(s) to test for temporal bias at. One or multiple of day or month, year. Can be abbreviated.

plot

a logical indicating whether to generate plots of spatial and temporal bias. See details for plot descriptions.

spatial.method

a character string, the method calculate the spatial bias statistic. One of; simple, convex_hull or core. See details.

centroid

a numeric vector of length two, specifying the centroid co-ordinate in the order of longitude then latitude. Only required if spatial.method = core. Default is mean of all occurrence record co-ordinates.

radius

a numeric value, the radial distance from the given centroid co-ordinate to measure spatial bias within. Only required if spatial.method = core. See details for more information. Default is mean distance of all co-ordinates from centroid.

prj

a character string, the coordinate reference system of occ.data co-ordinates. Default is "+proj=longlat +datum=WGS84".

Value

Returns list containing chi-squared and t-test results, and plots if specified.

Temporal bias To assess temporal sampling bias, the function returns a histogram plot

of the frequency distribution of records across the given time step specified by temporal.level (if plot = T). The observed frequency of sampling across the categorical time steps are compared to the distribution expected from random sampling, using a chi-squared test.

Spatial bias To assess spatial sampling bias, the function returns a scatterplot of the spatial

distribution of occurrence records to illustrate any spatial clustering (if plot = T). The average nearest neighbour distance of record co-ordinates is then compared to that of records randomly generated at same density using a t-test.

Spatial bias: methods * simple - generates the random points within a rectangle created using

the minimum and maximum longitude and latitude of occurrence co-ordinates.

convex_hull - generates the random points within the convex hull of occurrence record co-ordinates (i.e. the smallest convex set that contains all records).

core - generates the random points within specified circular area generated from a centroid point and radius. If these arguments ( centroid and radius) are not provided then centroid is calculated by averaging co-ordinates of all occurrence records, and radius is the mean distance away of all records from the centroid.

The core approach enables users to explore sampling bias in set areas of a species range. This may be valuable if species periphery-core relationships could lead to inaccurate inferences of sampling bias.

Computation time

As the spatial bias test involves the calculation of a distance matrix. To reduce computation time, it is recommended that only a representative sample of large occurrence datasets are input.

Examples

data(sample_explan_data)
bias_simple <- spatiotemp_bias(
occ.data = sample_explan_data,
temporal.level = c("day", "month", "year"),
spatial.method = "simple",
plot = FALSE
)
bias_core <-spatiotemp_bias(
occ.data = sample_explan_data,
temporal.level = c("day", "month", "year"),
spatial.method = "core",
plot = FALSE
)