Check species occurrence record formatting, completeness and validity.

Checks the occurrence record data frame contains the column names and classes required for dynamicSDM functions. Option to exclude records containing missing, duplicate or invalid co-ordinates or dates.

spatiotemp_check(
  occ.data,
  na.handle,
  duplicate.handle,
  coord.handle,
  date.handle,
  date.res = "day",
  coordclean = F,
  coordclean.species,
  coordclean.handle = "exclude",
  ...
)

Arguments

occ.data

a data frame, with columns for occurrence record co-ordinates and dates with column names as follows; record longitude as "x", latitude as "y", year as "year", month as "month", and day as "day".

na.handle

a character string, method for handling missing data (NA values) in record co-ordinates and dates. One of exclude or ignore: can be abbreviated. Default; exclude.

duplicate.handle

a character string, method for handling duplicate record co-ordinates or dates. One of exclude or ignore: can be abbreviated. Default; exclude.

coord.handle

a character string, method for handling invalid co-ordinates in record data. One of exclude or ignore: can be abbreviated. Default; exclude.

date.handle

a character string, method for handling invalid dates in record data. One of exclude or ignore: can be abbreviated. Default; exclude.

date.res

a character string, stating the temporal resolution to exclude dates by. One of year, month or day.

coordclean

a logical indicating whether to run function clean_coordinates from package CoordinateCleaner on occ.data. Default = FALSE.

coordclean.species

a character string or vector, specifying the name of the species that all of occ.data records belong to, or a character vector the length of nrow(occ.data) specifying which species each record belongs to. Required if coordclean = TRUE.

coordclean.handle

a character string, method for handling records flagged by CoordinateCleaner. One of exclude or report. Default: exclude.

...

Other arguments passed onto CoordinateCleaner.

Value

By default, returns occurrence record data frame, filtered to exclude records containing missing, duplicate or invalid data in record co-ordinates and dates.

Details

Record dates and co-ordinates are checked for validity using the following rules:

Dates must be real dates that could exist. For example, 50th February 2000 is not a valid date.

Co-ordinates must have longitude (x) values between -180 and 180, and latitude (y) values between -90 and 90 to be considered valid.

CoordinateCleaner compatibility

spatiotemp_check() acts as a helper function for compatibility with the R package CoordinateCleaner (Zizka et al., 2019), which offers a diversity of functions for checking the co-ordinates of occurrence records.

If coordclean = T, then coordclean.species must be provided to identify which species each record belonds to. If coordclean.handle = exclude then all occ.data records flagged by CoordinateCleaner::clean_coordinates() as potentially erroneous are removed in the returned data. If coordclean.handle = report, then the in-built report output by CoordinateCleaner::clean_coordinates()is returned. This contains logicals specifying the potentially erroneous records.

References

Zizka A, Silvestro D, Andermann T, Azevedo J, Duarte Ritter C, Edler D, Farooq H, Herdean A, Ariza M, Scharn R, Svanteson S, Wengstrom N, Zizka V, Antonelli A (2019). “CoordinateCleaner: standardized cleaning of occurrence records from biological collection databases.” Methods in Ecology and Evolution, -7. doi:10.1111/2041-210X.13152, R package version 2.0-20, https://github.com/ropensci/CoordinateCleaner.

Examples

data(sample_occ_data)
sample_occ_data<-convert_gbif(sample_occ_data)

nrow(sample_occ_data)
#> [1] 600

filtered<-spatiotemp_check(
 occ.data = sample_occ_data,
 coord.handle = "exclude",
 date.handle = "exclude",
 duplicate.handle = "exclude",
 na.handle = "exclude"
)
#> omitting any species records with coordinates containing NA
#> omitting any species records with dates containing NA
#> omitting any duplicate records
#> any records with invalid co-ordinates excluded
#> any records with invalid dates excluded
nrow(filtered)
#> [1] 311

if (FALSE) {
filtered_CC<-spatiotemp_check(
 occ.data = sample_occ_data,
 coord.handle = "exclude",
 date.handle = "exclude",
 duplicate.handle = "exclude",
 na.handle = "exclude",
 coordclean = T,
 coordclean.species = "quelea",
 coordclean.handle = "exclude"
)
nrow(filtered_CC)
}