Generate pseudo-absence record coordinates and dates

Function generates specified number of pseudo-absence record co-ordinates and dates either randomly or buffered in space and time.

spatiotemp_pseudoabs(
  spatial.method,
  temporal.method,
  occ.data,
  spatial.ext,
  temporal.ext,
  spatial.buffer,
  temporal.buffer,
  n.pseudoabs = 100,
  prj = "+proj=longlat +datum=WGS84"
)

Arguments

spatial.method

a character string, the spatial method for pseudo-absence generation. One of buffer or random: can be abbreviated.

temporal.method

a character string, the temporal method for pseudo-absence generation. One of buffer or random: can be abbreviated.

occ.data

optional; a data frame, with columns for occurrence record co-ordinates and dates with column names as follows; record longitude as "x", latitude as "y", year as "year", month as "month", and day as "day". Required if either temporal.method or spatial.method is buffer.

spatial.ext

the spatial extent to randomly generate pseudo-absences within. Object from which extent can be extracted of class Extent, RasterLayer, sf or polygon or numeric vector listing xmin, xmax, ymin and ymax in order. Required if spatial.method is random, and optionally used if buffer. See details.

temporal.ext

optional; a character vector, two dates in format "YYYY-MM-DD". First represents start of temporal extent and second represents end of temporal extent to randomly generate pseudo-absences dates within. Required if temporal.method is random, and optionally used if buffer. See details.

spatial.buffer

optional; a numeric value or vector, the radius/radii in metres to generate buffered pseudo-absence coordinates within. Only required if spatial.method is buffer. See details.

temporal.buffer

optional; a numeric value or vector, the period(s) in days to generate buffered pseudo-absence dates within. Only required if temporal.method is buffer . See details.

n.pseudoabs

optional; a numeric value, the number of pseudo-absence records to generate. Default; 100.

prj

a character string, the coordinate reference system of input occ.data co-ordinates Default is "+proj=longlat +datum=WGS84".

Value

Returns data frame of pseudo-absence coordinates and dates.

Spatial buffer

If spatial.method is buffer, then the pseudo-absence record co-ordinates are randomly generated in a buffered area defined either by

single numeric value for spatial.buffer - anywhere between the occurrence record and the circular distance surrounding this point (as specified in metres).

two numeric values for spatial.buffer - anywhere between the closest radius from the occurrence record and the furthest away radius (as specified in metres).

For example, if spatial.buffer = c(3000,10000), then pseudo-absence co-ordinates are randomly generated at least 3000m radius away from occurrence record co-ordinate but within 10000m radius. Whereas, if spatial.buffer = 10000, then pseudo-absence co-ordinates are randomly generated anywhere between 0m and 10000m radius from the occurrence record.

If spatial.ext is given too, then the generated pseudo-absences are not only constrained to the buffered area but also to this extent. For instance, if occurrence records are coastal, you may want to clip buffers to only terrestrial regions using a country polygon given in spatial.ext.

Spatial random

If spatial.method is random, then the pseudo-absence record co-ordinates are randomly generated across spatial.ext object given.

If spatial.ext is an sf polygon or multipolygon object, RasterLayer (mask if possible before input) or SpatialPolygonsDataFrame, then these shapes are used, instead of a simple rectangular extent. Therefore, inputting one of these objects will allow for more specific pseudo-absence generation.

For example, inputting a "SpatialPolygonsDataFrame" of specific countries will ensure co-ordinates are terrestrial, whereas an extent that encompasses these countries may result in the generation of pseudo-absence records in inappropriate areas, such as oceans or non-study-area countries.

Temporal buffer

If temporal.method is buffer, then pseudo-absence record dates are randomly generated between in a period defined by:

single numeric value for temporal.buffer - any date between the occurrence record date and the total number of days specified prior or post.

two numeric values for temporal.buffer - any date between the closest and furthers away number of days specified.

For example, if temporal.buffer = c(14,30), then pseudo-absence dates randomly generated at least 14 days from occurrence record dates but within 30 days. Whereas if temporal.buffer = 30, pseudo-absence dates are randomly generated anywhere between 0 and 30 days prior or post the occurrence record date.

If temporal.ext is given too, then the generated pseudo-absence dates are not only constrained to the buffer period but also to this temporal extent. For instance, an occurrence record recorded at the start of temporal.ext with 7 day buffer, may result in generated pseudo-absences outside of the temporal extent of the study.

Temporal random

If temporal.method is random, then pseudo-absence record dates are randomly generated within the two temporal.ext dates given.

Examples

data("sample_filt_data")
# \dontshow{
sample_filt_data<-sample_filt_data[1:10,]
# }

spatiotemp_pseudoabs(
 sample_filt_data,
 spatial.method = "random",
 temporal.method = "random",
 spatial.ext = c(20, 36, -35, -12),
 temporal.ext = c("2011-01-01", "2017-01-01")
)
#>            x         y year month day
#> 1   20.34479 -15.85394 2015     5  31
#> 2   27.04267 -23.43223 2011    10  24
#> 3   26.31505 -20.24687 2013     8  29
#> 4   31.57736 -14.93799 2011    11  13
#> 5   26.89067 -13.32659 2015     4  18
#> 6   31.07921 -22.10666 2012    12  29
#> 7   25.71924 -31.44756 2011    12   3
#> 8   35.52809 -17.53066 2013     2  10
#> 9   27.39533 -24.04886 2013     5  17
#> 10  29.99316 -15.51641 2011     9  30
#> 11  33.23898 -16.58121 2016     2  29
#> 12  26.10075 -24.55619 2013     3  13
#> 13  25.36888 -27.42885 2015     6   2
#> 14  25.86260 -18.47384 2013    10  30
#> 15  23.76395 -27.97756 2013     3  17
#> 16  28.10364 -14.24608 2013    12   5
#> 17  33.98056 -29.92855 2011     3   3
#> 18  31.56560 -24.93747 2014     6  17
#> 19  20.06596 -22.52764 2013    11   3
#> 20  31.33966 -30.17328 2013    11  27
#> 21  30.68607 -26.21322 2016     4  22
#> 22  22.09169 -27.04444 2014     7   2
#> 23  22.74380 -29.33472 2014    12  30
#> 24  33.62960 -31.85622 2012    12   3
#> 25  27.27175 -18.36916 2015    11  14
#> 26  25.86580 -12.78727 2013     1   8
#> 27  30.67278 -19.82064 2011     5  31
#> 28  35.72962 -30.15408 2016     7  28
#> 29  22.05842 -29.23646 2012     5  13
#> 30  29.54822 -26.60455 2014    12  27
#> 31  31.55124 -12.18592 2012     2   6
#> 32  27.96233 -16.95293 2015    10  19
#> 33  22.17734 -25.82832 2012     3   9
#> 34  30.20306 -13.54224 2012     9  28
#> 35  31.85519 -33.53205 2016     5  25
#> 36  32.44431 -12.78043 2016     7  25
#> 37  35.12939 -14.09307 2014    10  13
#> 38  30.99987 -33.28570 2015    12  14
#> 39  20.63257 -23.65815 2011    10   5
#> 40  24.90211 -30.25439 2014     9  25
#> 41  29.27979 -15.39810 2012     1   2
#> 42  35.89593 -25.37539 2015    11  19
#> 43  31.76660 -28.71068 2011     6  10
#> 44  32.62131 -13.95970 2012     5  31
#> 45  20.81119 -24.11962 2016     7  24
#> 46  25.65696 -16.78178 2015     7   1
#> 47  23.26066 -20.10614 2015     2  20
#> 48  25.68658 -28.01891 2012     9   7
#> 49  34.67699 -33.45029 2016     4  21
#> 50  32.17167 -31.77808 2012     8  20
#> 51  23.33957 -31.84123 2014     8  28
#> 52  32.15629 -17.35408 2016    12  26
#> 53  32.50607 -15.63657 2013     4  24
#> 54  25.11207 -18.97995 2016     8  18
#> 55  27.32810 -13.39739 2013    10  28
#> 56  20.10484 -24.01252 2016     7   8
#> 57  34.86137 -23.95423 2012     6  27
#> 58  20.89013 -20.25889 2013    12  24
#> 59  22.66652 -12.65491 2013     8   8
#> 60  28.25555 -25.46435 2016    11  28
#> 61  31.79040 -21.03380 2012     5  26
#> 62  33.92159 -25.74637 2011     7  15
#> 63  28.60392 -25.56205 2014     1   9
#> 64  27.72333 -19.30884 2015    11  21
#> 65  35.92847 -30.16155 2013    10  15
#> 66  25.39103 -25.67682 2014    11  24
#> 67  20.07073 -26.13673 2013     8  17
#> 68  29.25965 -19.35973 2014     7  16
#> 69  22.18376 -20.88660 2015     8   8
#> 70  34.03625 -12.53967 2016     9  11
#> 71  26.82997 -26.87450 2011    11  11
#> 72  24.33557 -24.82834 2014    10   5
#> 73  33.64901 -17.26452 2012     8  28
#> 74  29.63395 -21.35550 2013     2  11
#> 75  22.42260 -18.57322 2016     6   3
#> 76  21.23525 -24.73771 2016     2   3
#> 77  28.12428 -15.32108 2015     7  12
#> 78  35.22907 -18.25455 2016     9  10
#> 79  20.11957 -20.66331 2015     2   9
#> 80  20.97178 -22.68311 2011     3   2
#> 81  31.59962 -29.42928 2012     9  24
#> 82  24.14658 -24.70258 2015     8  17
#> 83  34.14046 -32.55492 2012     3  25
#> 84  34.71234 -27.46361 2015     7  25
#> 85  31.89632 -34.89440 2016    11  13
#> 86  21.40634 -28.75952 2012     2   2
#> 87  25.23312 -22.84282 2015     8  13
#> 88  28.39317 -12.89981 2012     1  29
#> 89  23.55804 -33.92297 2011     5  27
#> 90  34.16874 -32.33085 2013     5  20
#> 91  28.83216 -30.60350 2012     6  19
#> 92  30.76915 -17.45623 2016     4  19
#> 93  34.78126 -30.62193 2014     4   2
#> 94  28.01880 -22.21129 2013    11  11
#> 95  35.10975 -20.25760 2013     7  23
#> 96  35.25127 -12.12245 2011     7  31
#> 97  35.75556 -28.68097 2011     6  27
#> 98  26.59624 -12.26403 2015     1   7
#> 99  33.40604 -16.81812 2013     3  15
#> 100 30.96813 -33.57564 2012    11  14