Thin species occurrence records by spatial and temporal proximity.

Thins species occurrence records that are within minimum spatial and temporal distance apart.

spatiotemp_thin(
  occ.data,
  temporal.method,
  temporal.dist,
  spatial.split.degrees,
  spatial.dist,
  iterations = 100
)

Arguments

occ.data

a data frame, with columns for occurrence record co-ordinates and dates with column names as follows; record longitude as "x", latitude as "y", year as "year", month as "month", and day as "day".

temporal.method

a character string, the method to calculate temporal distance between records. One of DOY or day. See details for more information.

temporal.dist

a numeric value, the temporal buffer in days to thin records by.

spatial.split.degrees

a numeric value, the grid cell resolution in degrees to split occurrence records by before temporal thinning.

spatial.dist

a numeric value, the spatial buffer distances in metres to thin records by.

iterations

a numeric value, the number of iterations to randomly thin occurrence records by. Default; 100.

Value

Returns data frame of occurrence records thinned by specified temporal and spatial distance.

Overview

spatiotemp_thin() calculates the temporal distance between occurrence records in given area and excludes records below minimum temporal distance apart. Then calculates spatial distance between all occurrence records and filters records below minimum spatial distance apart. Thus, spatiotemp_thin() offers spatial and temporal thinning, improving functionality for dynamic species distribution modelling compared to existing spThin package functions (Aiello-Lammens et al., 2015).

Temporal thinning methods

For temporal thinning, the function first splits occurrence records into grid cells of given size in degrees (set by spatial.split.degrees). This is to prevent spatially distant but temporally close records from being excluded. For each grid cell, all records within the cell are temporally thinned. This process works by removing records that are within given temporal distance (temporal.dist) from each other by randomly selecting one of the two. This iterates through until no records are within the given temporal distance of each other in each grid cell, following a similar algorithm to spThin (Aiello-Lammens et al., 2015).

Two methods exist for measuring the temporal distance between occurrence records.

doy - calculates the minimum days apart within the annual cycle

day uses the absolute number of days.

For instance, two dates “2010-01-05” and “2012-12-05” can be calculated as either 1065 absolute days apart, or within the annual cycle these dates represent day 5 and day 339 of the year, and are 31 days apart. Therefore, thinning by 40 days using the DOY method would remove one of these records, but using the day method would not. The chosen temporal.method will depend upon whether bias towards a point within the annual cycle is more likely to bias models than bias towards a given point in linear time.

Spatial thinning

Following temporal thinning, spatial thinning occurs across entire dataset. The spatial distance between each record is calculated, and records within the given spatial distance (spatial.dist) from each other are excluded by randomly selecting one of these. This iterates through until no records are with the given spatial distances of each other across entire dataset, again following a similar algorithm to spThin (Aiello-Lammens et al., 2015).

As random selection could alter the total number of occurrence records remaining in the occurrence record dataset, this process is iterated through a specified number of times and the thinned data frame with the highest number of records remaining is returned.

References

Aiello-Lammens, M. E., Boria, R. A., Radosavljevic, A., Vilela, B. & Anderson, R. P. 2015. spThin: an R package for spatial thinning of species occurrence records for use in ecological niche models. Ecography, 38, 541-545.

Examples

data("sample_filt_data")

spatiotemp_thin(
 occ.data = sample_filt_data,
 temporal.method = "day",
 temporal.dist = 100,
 spatial.split.degrees = 3,
 spatial.dist = 100000,
 iterations = 5
)
#>            x         y year month day decimalLatitude decimalLongitude
#> 13  15.48417 -19.35105 2016     7   3       -19.35105         15.48417
#> 16  16.93718 -18.76557 2018     7  11       -18.76557         16.93718
#> 111 17.86674 -22.55155 2016     5   2       -22.55155         17.86674
#> 18  22.16667 -26.53333 2005     5  18       -26.53333         22.16667
#> 22  24.62272 -26.54049 2017    12   1       -26.54049         24.62272
#> 32  23.45594 -28.37425 2005    10  24       -28.37425         23.45594
#> 81  24.84293 -28.10646 2018     9  11       -28.10646         24.84293
#> 33  24.37663 -33.29126 2016     1  17       -33.29126         24.37663
#> 73  25.45792 -31.37458 2014     1   2       -31.37458         25.45792
#> 53  25.13545 -17.81830 2015     8  19       -17.81830         25.13545
#> 101 26.29125 -18.45792 2013     4  23       -18.45792         26.29125
#> 181 26.73333 -26.36667 2009     5  25       -26.36667         26.73333
#> 40  27.70792 -26.87458 2018     7  12       -26.87458         27.70792
#> 56  25.51798 -28.83252 2016    12  18       -28.83252         25.51798
#> 182 27.37582 -28.95870 2007     3  18       -28.95870         27.37582
#> 117 26.50205 -24.68161 2016     8  26       -24.68161         26.50205
#> 118 27.41655 -21.34998 2004     9   1       -21.34998         27.41655
#> 401 28.58333 -25.95000 2005     7  12       -25.95000         28.58333
#> 522 29.37353 -25.37604 2018     9  11       -25.37604         29.37353
#> 57  29.45901 -27.20872 2016     5  20       -27.20872         29.45901
#> 641 30.74723 -27.03281 2016    10  20       -27.03281         30.74723
#> 119 28.31454 -28.24909 2018     2   8       -28.24909         28.31454
#> 67  29.95792 -29.37458 2013    10  26       -29.37458         29.95792
#> 142 30.04271 -28.29271 2005     7  23       -28.29271         30.04271
#> 120 28.31667 -24.90000 2007     1   9       -24.90000         28.31667
#> 85  30.37642 -24.79113 2018     5  18       -24.79113         30.37642
#> 212 28.68333 -20.18333 2007     1   9       -20.18333         28.68333
#> 36  29.83333 -19.50000 2007    10  11       -19.50000         29.83333
#> 68  31.12536 -22.28966 2006     3  22       -22.28966         31.12536
#> 191 31.46001 -23.45804 2009     2  27       -23.45804         31.46001
#> 243 31.79111 -24.37245 2016     7   3       -24.37245         31.79111
#> 310 31.16767 -17.66439 2008     2  23       -17.66439         31.16767
#> 87  32.62458 -18.95792 2015    12  16       -18.95792         32.62458
#> 311 31.79160 -25.62652 2009     7   1       -25.62652         31.79160
#> 96  31.88483 -27.40119 2016    12  27       -27.40119         31.88483
#> 214 33.58406 -10.83137 2018     5  12       -10.83137         33.58406
#> 127 34.36077 -18.96188 2017     8  20       -18.96188         34.36077
#>     occurrenceStatus     source species SAMP_EFFORT REL_SAMP_EFFORT split2
#> 13            absent dynamicSDM     RBQ           0       0.1428571     32
#> 16            absent dynamicSDM     RBQ           0       0.1428571     23
#> 111           absent dynamicSDM     RBQ           0       0.1428571     43
#> 18           present   gbifdata     RBQ           0       0.1428571     55
#> 22            absent dynamicSDM     RBQ           2       0.4285714     55
#> 32            absent dynamicSDM     RBQ           0       0.1428571     65
#> 81            absent dynamicSDM     RBQ           0       0.1428571     65
#> 33            absent dynamicSDM     RBQ           0       0.1428571     75
#> 73           present   gbifdata     RBQ           0       0.1428571     76
#> 53            absent dynamicSDM     RBQ           0       0.1428571     26
#> 101          present   gbifdata     RBQ           0       0.1428571     26
#> 181          present   gbifdata     RBQ           0       0.1428571     56
#> 40           present   gbifdata     RBQ           0       0.1428571     56
#> 56            absent dynamicSDM     RBQ           0       0.1428571     66
#> 182           absent dynamicSDM     RBQ           0       0.1428571     66
#> 117           absent dynamicSDM     RBQ           0       0.1428571     46
#> 118          present   gbifdata     RBQ           0       0.1428571     36
#> 401          present   gbifdata     RBQ           0       0.1428571     57
#> 522           absent dynamicSDM     RBQ           0       0.1428571     57
#> 57            absent dynamicSDM     RBQ           0       0.1428571     57
#> 641          present   gbifdata     RBQ           0       0.1428571     57
#> 119           absent dynamicSDM     RBQ           0       0.1428571     67
#> 67           present   gbifdata     RBQ           0       0.1428571     67
#> 142           absent dynamicSDM     RBQ           0       0.1428571     67
#> 120          present   gbifdata     RBQ           0       0.1428571     47
#> 85            absent dynamicSDM     RBQ           0       0.1428571     47
#> 212          present   gbifdata     RBQ           0       0.1428571     37
#> 36           present   gbifdata     RBQ           0       0.1428571     37
#> 68            absent dynamicSDM     RBQ           0       0.1428571     48
#> 191           absent dynamicSDM     RBQ           0       0.1428571     48
#> 243           absent dynamicSDM     RBQ           0       0.1428571     48
#> 310           absent dynamicSDM     RBQ           0       0.1428571     28
#> 87           present   gbifdata     RBQ           0       0.1428571     28
#> 311           absent dynamicSDM     RBQ           1       0.2857143     58
#> 96            absent dynamicSDM     RBQ           0       0.1428571     58
#> 214           absent dynamicSDM     RBQ           0       0.1428571      8
#> 127          present   gbifdata     RBQ           0       0.1428571     29