Calculate sampling effort across spatial and temporal buffer from species occurrence records

Calculates the total number of sampling events across a given spatial and temporal buffer from each occurrence record’s co-ordinate and date.

spatiotemp_weights(
  occ.data,
  samp.events,
  spatial.dist = 0,
  temporal.dist = 0,
  prj = "+proj=longlat +datum=WGS84"
)

Arguments

occ.data

a data frame, with columns for occurrence record co-ordinates and dates with column names as follows; record longitude as "x", latitude as "y", year as "year", month as "month", and day as "day".

samp.events

a data.frame, sampling events with column names as follows; record longitude as "x", latitude as "y", year as "year", month as "month", and day as "day".

spatial.dist

a numeric value, the spatial distance in metres representing the radius from occurrence record co-ordinate to sum sampling events across.

temporal.dist

a numeric value, the temporal distance in days, representing the period before and after the occurrence record date to sum sampling events across.

prj

a character string, the coordinate reference system of input occ.data co-ordinates Default is "+proj=longlat +datum=WGS84".

Value

Returns input occurrence record data frame with additional columns for sampling effort "SAMP_EFFORT" and relative sampling effort "REL_SAMP_EFFORT".

Details

For each occurrence record, this function calculates the total number of sampling events within given radius (spatial.dist) from each record co-ordinate and days (temporal.dist) both prior and post record date.

In addition to total sampling events, the function also calculates relative sampling effort, scaling from 0 (least sampled) to 1 (most sampled).

Output could be used to calculate model weights to correct spatial and temporal biases in occurrence record collections.

Examples

data("sample_explan_data")
data("sample_events_data")
# \dontshow{
sample_explan_data<-sample_explan_data[1:5,]
# }

spatiotemp_weights(
 occ.data = sample_explan_data,
 samp.events = sample_events_data,
 spatial.dist = 200000,
 temporal.dist = 20
)
#>          x         y year month day decimalLatitude decimalLongitude
#> 1 14.36667 -19.31667 2005     5  18       -19.31667         14.36667
#> 3 15.21667 -19.25000 2005     5  18       -19.25000         15.21667
#> 5 15.48090 -19.34948 2005     8   5       -19.34948         15.48090
#> 4 15.41667 -19.36667 2006     1  16       -19.36667         15.41667
#> 2 15.21503 -19.24880 2016    12  27       -19.24880         15.21503
#>   occurrenceStatus     source species SAMP_EFFORT REL_SAMP_EFFORT unique.ID.DYN
#> 1          present   gbifdata     RBQ           0       0.1428571             9
#> 3          present   gbifdata     RBQ           0       0.1428571            11
#> 5           absent dynamicSDM     RBQ           0       0.1428571            30
#> 4          present   gbifdata     RBQ           0       0.1428571            47
#> 2           absent dynamicSDM     RBQ           0       0.1428571           259
#>   eight_sum_prec year_sum_prec grass_crop_percentage presence.absence
#> 1       10.36304      210.8262                     0                1
#> 3       12.15311      271.8394                   600                1
#> 5        0.00000      292.9795                   931                0
#> 4       62.50370      246.0937                     0                1
#> 2       29.85218      198.2361                     0                0
#>   SAMP_EFFORT REL_SAMP_EFFORT
#> 1           0             Inf
#> 3           0             Inf
#> 5           0             Inf
#> 4           0             Inf
#> 2           0             Inf